Add a config for coffeescript
This doesn't set up tests yet, and with the way the project is set up, it ends
up running fewer tests that pass, so I added a flag to just skip tests if we
know that they aren't set up right.

Also, we only support CoffeeScript 1.10.0 at the moment, so I added a way to
check out a custom branch when cloning the repo.

This is also the first repo with files that we know we can't convert. We now
detect if there were failures and run decaffeinate on just the successful files
if so.