Add a default config to simplify the per-project config (#5)
At least so far, every project has needed to install five packages and copy two
config files for babel and eslint, so those can all be included with a single
flag in the config file. This also gets rid of the config-files directory, since
I don't need it anymore, but it may come back if it comes up more.

Also exit 1 on failure and make travis run through the hubot conversion.

Also do a bunch of stuff to get that travis build set up.