Add a --publish option
By adding `--publish` to the result, the tool will now add the decaffeinate fork
as a remote and force-push a decaffeinate branch with the results. It also
writes a summary of the conversion results to the README. Currently the summary
isn't particularly pretty, but hopefully it can have more useful information in
the future, like number of lines converted.