Clear node_modules before the atom build
When I run the atom build locally, it complains about version mismatches in
node_modules since the build script wants to install that itself. It's a little
non-trivial to change decaffeinate-examples to always leave that up to the
project, so to get things working I'll just get rid of the directory before
running tests.