First pass at CoffeeScript test config
For now, this doesn't even use the new version of the compiler, it just runs all
of the tests against the old compiler. There are a bunch of errors, so seems
good to work through those as a starting point. Since the tests are incomplete,
I made it so they always fail.