Replace `run` with a simpler version (#32)
The previous `run` was fancy in that it tried to produce live output and also
make stdout/stderr available to the JS side afterward, but this added a lot of
complexity, e.g. using pty so the child process will treat us like a terminal,
and the stdout/stderr ended up never being used. The atom tests are failing with
no output, so this should at least reduce the number of variables in figuring
out what's going wrong.