Avoid running decaffeinate multiple times for each file
We don't need to verify during convert since there's already a check step
earlier. Also, if we expect all files to convert, we can just skip the check
step and let the build fail in an uglier way. If it does fail, we'll have to
change `expectConversionSuccess` to false or fix the failure.