Add retry logic if there are failures after skipping the check command
If decaffeinate introduces a regression that causes crashing bugs (like just
happened), then the behavior was for those builds to fail, but the dashboard
wouldn't update, since `expectConversionSuccess` skips the robust error
reporting code path. Recovering from a failed decaffeinate operation is hard
since we don't know what state the git repo is in, so instead, we retry the
whole operation on a fresh repo if we discover that there's a crash during
decaffeinate. This should cause it to properly update the results badge to
indicate the test failures.