Patch vimium to fix the build (#103)
Vimium just landed a change that set a class to the empty object as a dummy
value for test purposes. This works in CoffeeScript, but you can't extend the
empty object in JS, so I just changed it to `class {}` instead. It's probably
fair to call this a limitation of decaffeinate and document it in the list of
known limitations.