Avoid initial npm install before running decaffeinate
This was causing issues for codecombat, and generally it seems nicer to only
install once rather than before and after. This might cause some projects to
break, but I can iterate on those, and the ones I tested seem fine.

Also change CoffeeScript config to expect the test suite to pass.